refactor: ensure delete not found throws

+++ src/services/DeleteTransactionService.ts
@@ -9,6 +9,10 @@ interface Request {
 class DeleteTransactionService {
   public async execute({ id }: Request): Promise<void> {
     const transactionRepository = getCustomRepository(TransactionRepository);
+    const transaction = await transactionRepository.findOne({ id });
+    if (!transaction) {
+      throw new AppError('Transaction not found', 404);
+    }
     await transactionRepository.delete({ id });
   }
 }
